move the schur tests into test_mb.py
